Move auth specs from user module to the authentication module
