+++ modules/authentication/server-scala/src/test/scala/GoogleAuthSpec.scala
@@ -4,36 +4,35 @@ import akka.http.scaladsl.testkit.RouteTestTimeout
 import akka.testkit.TestDuration
 import com.github.scribejava.core.model.{OAuth2AccessToken, OAuthRequest, Response}
 import com.github.scribejava.core.oauth.OAuth20Service
-import repositories.auth.LinkedinAuthRepository
-import services.ExternalApiService
-import routes.auth.LinkedinAuthController
 import modules.jwt.model.JwtContent
 import modules.jwt.service.JwtAuthService
-import repositories.UserRepository
+import repositories.{GoogleAuthRepository, UserRepository}
+import routes.GoogleAuthController
+import services.ExternalApiService
 
 import scala.concurrent.ExecutionContext
 import scala.concurrent.duration._
 
-class LinkedinAuthSpec extends UserHelper {
+class GoogleAuthSpec extends UserTestHelper {
   implicit val timeout: RouteTestTimeout = RouteTestTimeout(10.seconds.dilated)
   val executionContext: ExecutionContext = inject[ExecutionContext]
 
   val userRepository: UserRepository = inject[UserRepository]
-  val authRepository: LinkedinAuthRepository = inject[LinkedinAuthRepository]
+  val authRepository: GoogleAuthRepository = inject[GoogleAuthRepository]
   val externalApiService: ExternalApiService = inject[ExternalApiService]
   val jwtAuthService: JwtAuthService[JwtContent] = inject[JwtAuthService[JwtContent]]
 
   val oAuth2ServiceMock: OAuth20Service = stub[OAuth20Service]
   val responseMock: Response = stub[Response]
 
-  val authController = new LinkedinAuthController(oAuth2ServiceMock, externalApiService, userRepository, authRepository, jwtAuthService)(executionContext)
+  val authController = new GoogleAuthController(oAuth2ServiceMock, externalApiService, userRepository, authRepository, jwtAuthService)(executionContext)
   val authRoutes: Route = authController.routes
 
-  "LinkedinAuthController" must {
-    "redirect to linkedin auth page" in {
+  "GoogleAuthController" must {
+    "redirect to google auth page" in {
       (() => oAuth2ServiceMock.getAuthorizationUrl).when.returns("localhostTest")
 
-      Get("/auth/linkedin") ~> authRoutes ~> check {
+      Get("/auth/google") ~> authRoutes ~> check {
         status shouldBe StatusCodes.Found
         status.isRedirection() shouldBe true
         responseAs[String] should include("localhostTest")
@@ -46,14 +45,14 @@ class LinkedinAuthSpec extends UserHelper {
       (() => responseMock.getBody).when.returns(
         """
           |{
-          |   "id":"test",
-          |   "email-address":"test@test.com",
-          |   "formatted-name":"testName"
+          |   "id":"testId",
+          |   "email":"test@test.com",
+          |   "name":"testName"
           |}
         """.stripMargin)
       ((request: OAuthRequest) => oAuth2ServiceMock.execute(request)).when(*).returns(responseMock)
 
-      Get("/auth/linkedin/callback?code=test") ~> authRoutes ~> check {
+      Get("/auth/google/callback?code=test") ~> authRoutes ~> check {
         status shouldBe StatusCodes.Found
         status.isRedirection() shouldBe true
         responseAs[String] should include("/profile")
@@ -70,14 +69,14 @@ class LinkedinAuthSpec extends UserHelper {
       (() => responseMock.getBody).when.returns(
         """
           |{
-          |   "id":"test",
-          |   "email-address":"test@test.com",
-          |   "formatted-name":"testName"
+          |   "id":"testId",
+          |   "email":"test@test.com",
+          |   "name":"testName"
           |}
         """.stripMargin)
       ((request: OAuthRequest) => oAuth2ServiceMock.execute(request)).when(*).returns(responseMock)
 
-      Get("/auth/linkedin/callback?state=test&code=test") ~> authRoutes ~> check {
+      Get("/auth/google/callback?state=test&code=test") ~> authRoutes ~> check {
         status shouldBe StatusCodes.Found
         status.isRedirection() shouldBe true
         responseAs[String] should include("test?data=")
@@ -91,7 +90,7 @@ class LinkedinAuthSpec extends UserHelper {
     }
 
     "redirect to login page if an error was capture" in {
-      Get("/auth/linkedin/callback?code=test") ~> authRoutes ~> check {
+      Get("/auth/google/callback?code=test") ~> authRoutes ~> check {
         status shouldBe StatusCodes.Found
         status.isRedirection() shouldBe true
         responseAs[String] should include("/login")
