+++ modules/authentication/server-scala/src/test/scala/LinkedinAuthSpec.scala
@@ -4,37 +4,35 @@ import akka.http.scaladsl.testkit.RouteTestTimeout
 import akka.testkit.TestDuration
 import com.github.scribejava.core.model.{OAuth2AccessToken, OAuthRequest, Response}
 import com.github.scribejava.core.oauth.OAuth20Service
-import repositories.auth.GithubAuthRepository
-import services.ExternalApiService
-import routes.auth.GithubAuthController
 import modules.jwt.model.JwtContent
 import modules.jwt.service.JwtAuthService
-import repositories.UserRepository
+import repositories.{LinkedinAuthRepository, UserRepository}
+import routes.LinkedinAuthController
+import services.ExternalApiService
 
 import scala.concurrent.ExecutionContext
 import scala.concurrent.duration._
 
-
-class GithubAuthSpec extends UserHelper {
+class LinkedinAuthSpec extends UserTestHelper {
   implicit val timeout: RouteTestTimeout = RouteTestTimeout(10.seconds.dilated)
   val executionContext: ExecutionContext = inject[ExecutionContext]
 
   val userRepository: UserRepository = inject[UserRepository]
-  val authRepository: GithubAuthRepository = inject[GithubAuthRepository]
+  val authRepository: LinkedinAuthRepository = inject[LinkedinAuthRepository]
   val externalApiService: ExternalApiService = inject[ExternalApiService]
   val jwtAuthService: JwtAuthService[JwtContent] = inject[JwtAuthService[JwtContent]]
 
   val oAuth2ServiceMock: OAuth20Service = stub[OAuth20Service]
   val responseMock: Response = stub[Response]
 
-  val authController = new GithubAuthController(oAuth2ServiceMock, externalApiService, userRepository, authRepository, jwtAuthService)(executionContext)
+  val authController = new LinkedinAuthController(oAuth2ServiceMock, externalApiService, userRepository, authRepository, jwtAuthService)(executionContext)
   val authRoutes: Route = authController.routes
 
-  "GithubAuthController" must {
-    "redirect to github auth page" in {
+  "LinkedinAuthController" must {
+    "redirect to linkedin auth page" in {
       (() => oAuth2ServiceMock.getAuthorizationUrl).when.returns("localhostTest")
 
-      Get("/auth/github") ~> authRoutes ~> check {
+      Get("/auth/linkedin") ~> authRoutes ~> check {
         status shouldBe StatusCodes.Found
         status.isRedirection() shouldBe true
         responseAs[String] should include("localhostTest")
@@ -47,14 +45,14 @@ class GithubAuthSpec extends UserHelper {
       (() => responseMock.getBody).when.returns(
         """
           |{
-          |   "id":1,
-          |   "email":"test@test.com",
-          |   "name":"testName"
+          |   "id":"test",
+          |   "email-address":"test@test.com",
+          |   "formatted-name":"testName"
           |}
         """.stripMargin)
       ((request: OAuthRequest) => oAuth2ServiceMock.execute(request)).when(*).returns(responseMock)
 
-      Get("/auth/github/callback?code=test") ~> authRoutes ~> check {
+      Get("/auth/linkedin/callback?code=test") ~> authRoutes ~> check {
         status shouldBe StatusCodes.Found
         status.isRedirection() shouldBe true
         responseAs[String] should include("/profile")
@@ -71,14 +69,14 @@ class GithubAuthSpec extends UserHelper {
       (() => responseMock.getBody).when.returns(
         """
           |{
-          |   "id":1,
-          |   "email":"test@test.com",
-          |   "name":"testName"
+          |   "id":"test",
+          |   "email-address":"test@test.com",
+          |   "formatted-name":"testName"
           |}
         """.stripMargin)
       ((request: OAuthRequest) => oAuth2ServiceMock.execute(request)).when(*).returns(responseMock)
 
-      Get("/auth/github/callback?state=test&code=test") ~> authRoutes ~> check {
+      Get("/auth/linkedin/callback?state=test&code=test") ~> authRoutes ~> check {
         status shouldBe StatusCodes.Found
         status.isRedirection() shouldBe true
         responseAs[String] should include("test?data=")
@@ -92,7 +90,7 @@ class GithubAuthSpec extends UserHelper {
     }
 
     "redirect to login page if an error was capture" in {
-      Get("/auth/github/callback?code=test") ~> authRoutes ~> check {
+      Get("/auth/linkedin/callback?code=test") ~> authRoutes ~> check {
         status shouldBe StatusCodes.Found
         status.isRedirection() shouldBe true
         responseAs[String] should include("/login")
